requirements.txt
flask==2.0.1
gunicorn==20.1.0
supabase==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
psycopg2-binary==2.9.9
python-dotenv==0.19.0
pillow==11.2.1 